Protect against cross-linked CDJs.
Trying to borrow a player's device ID for a metadata query only works
if that player hasn't linked to the metadata player's database. This
change watches for that situation and logs an error if it would not be
safe to try to make the query. It also notices if the virtual CDJ is
using a valid player number, and if it is, always uses that as as the
ID of the device that wants to make the query instead of trying to
borrow an actual player's ID.