Protect against reading messages with missing binary fields.
It turns out that for huge tracks, there is no beat grid, and the
request comes back with a zero number field for the blob length, and
then just omits the blob field (grr!). This code tries to protect
against that case in general when reading messages with blob types.
What a truly poor protocol, though! They should always write up to the
length part of the blob header when the field is declared.