Don't hold double the art cache while resizing.
Copying all the entries into the new cache before getting rid of the
old means we need twice as much memory, even briefly. Now we get rid
of them as we go, so we have no extra memory demand.

At the same time, added a sanity check to the size the user asked for.