Improve concurrency, see #8
Much of what was suggested in that issue was already implemented in
the refactoring to the new data package, but this goes even further,
using Atomic reference classes to get rid of most of the remaining
synchronized methods. More work is in progress to adopt a concurrent
linked hashmap, and use copies to avoid issues with concurrent
modification of collections without synchronization.