Updating to conform to KSY Style Guide
Thanks to @GreyCat for calling my attention to this, it is hugely
helpful, not only in terms of improving the clarity of naming and
organization, but also in helping me see that I can (and should) use
fewer types to simplify the parse results and make them easier to work
with as well as more efficient. There is more to come on that front.