Restructure row index to use arrays.
Building on the successful linked-list implementation, with the help
of the simplification and cleanup just performed (and some swim-time
thinking), this is now working exactly the way I hoped it could, and
even presents the rows in proper logical order, while traversing the
reverse index from the back of the page to the heap. Woo hoo!