Revert "Switch to using sub-streams for positioning."
This reverts commit 4daf3f3df2272b2d37991d4fee79e39cd48ca084.

Although it led to cleaner-looking definitions in the `.ksv` file,
inspection of the generated Java class showed that it would create far
too many copies of a large fraction of the entire page while iterating
over rows in a table. So, roll back to the more primitive and explicit
approach until `ksc` can be more efficient about reusing byte buffers.