Always draw hot cue markers last
When loading from Crate Digger, we are not guaranteed to get the
memory cue entries before the hot cue entries, so it can be impossible
to see which are both in the waveform view components. This restores
the previous behavior where a hot cue is always drawn in green even if
there is also a memory cue in the same spot.