Reworking to support multiple playback positions.
When drawing waveforms in a show file, there might be more than one
player playing the track at a given time, so the component needs to be
able to model and visualize this. This is a first stab at changing the
preview component to support this new mode of operation.