Fix TimeFinder treatment of tracks reaching the end.
This is not nearly as big an issue as the infinite-loop bug, and did not
affect anything other than the interpolated playback time, but the code
that was intended to check if the track had stopped playing and stop
interpolating was instead only triggering if the track was paused. Since
reaching the end stops in a way different than pausing, the counter
would keep going in that situation, and the waveform detail component
would drift off into emptiness.