Delay between repeated attempts to fetch player files
Per @Kevinnns, retrying when he inserted media was happening
immediately, and failing each time, because perhaps the export
was not yet ready. Now we wait 2s before the second try, and an
additional 4s before the third try. If configured to allow more
retries, the maximum wait between them is 6s.