Stop TimeFinder being fooled by beats from pre-nexus players.
It now knows that it simply never gets a track position from them.
Also added a flag on the DeviceUpdate class that indicates when it
was received from a pre-nexus player (CDJ-900 or CDJ-2000) so BLT
and other programs can indicate to the user that track position
information won't be available.