Revert "Use weak references for listener lists, closes #85"
This reverts commit 091962ec9dbc8988f6dfeaaedea62d7bc889bf6c.

It caused delivery of events to become unreliable, so I am backing off
from this idea for now, and returning to the model that objects are
responsible for unregistering themselves from Beat Link if they want
to be eligible for garbage collection.

I will follow up with another commit shortly that reinstates a couple
of small unrelated fixes that were part of the reverted one.