Add plumbing so SignatureFinder can always get RGB waveforms
Almost everything is now in place, and it works fine when changing
the preferred waveform style already, as long as the RGB version
has been requested at some point. The only remaining task is to
ensure that all formats requested by any registered listener are
retrieved automatically, which will require adding an extension
to the MetadataProvider interface (an optional one, to retain
backwards compatibility), allowing specific, non-preferred formats
to be retrieved on demand.