Stabilize track signatures
Change the SignatureFinder to always use the RGB waveform detail
when calculating the track hash, so that changing the preferred
display style has no effect. Use the low-level analysis tag finder
to do this, so we can rip out all the complexity I had been adding
to the WaveformFinder to support tracking multiple styles.